+++ src/discljord/http.clj
@@ -2,22 +2,17 @@
   (:require [clojure.spec.alpha :as s]
             [discljord.specs :as ds]))
 
-(defn append-api-suffix
-  "Takes a url and appends GET parameters for version of the Discord
-  API, and the default encoding."
-  [url]
-  (str url "?v=6&encoding=json"))
-(s/fdef append-api-suffix
-  :args (s/cat :url ::ds/url)
-  :ret ::ds/url)
-
 (defn api-url
   "Takes an endpoint from Discord's API (with a starting '/'), and returns
   a URL for that Discord endpoint.
 
   For example: (api-url (str \"/channels/\" channel-id))"
-  [gateway]
-  (append-api-suffix (str "https://discordapp.com/api" gateway)))
+  [endpoint]
+  (str "https://discordapp.com/api/v6" endpoint))
 (s/fdef api-url
   :args (s/cat :url ::ds/url)
   :ret ::ds/url)
+
+(def gateway-url
+  "URL used to retrieve the gateway for the bot."
+  "https://discordapp.com/api/gateway/bot?v=6&encoding=json")